refactor: modification to models and init files, tests: writing tests to fit modified endpoints
